Stop clearing glossary's other terms on adding from "More Details" section
Previously, when adding terms from a Search page's display of all terms
in a glossary, any terms that were already in the cart but were not
checked were removed from the cart when the new terms were added. We've
found most users don't expect this behavior and see it as a frustrating
bug.