Double check columns hash before making changes in Collibra
Occasionally (once every few dozens of thousands of calls) the API
reading tables' columns gave an error. This resulted in the
database-syncing job mistakenly removing a table from Collibra only to
add it again the next time the job ran. Now if we see our hash of the
database information has changed, we'll fetch that data a second time to
be sure it's accurate.